fix(pkg): fixed published build issues
Version 1.0.0 previously caused crashes on React Native due to the way this is package during NPM publishes.

This was because I had installed "react", "next", and "react-native" as main dependencies during testing and forgot to remove it for production release.

Next has been moved to devDependencies alongside peerDependencies. React and React Native are only installed as types in devDependencies. This way, release builds don't actually cause multiple runtimes to be installed.