Support for i18n locales!
Per usual, this works just like NextJS. An `i18n` object is passed as a configuration for Pilot to use.

It follows the same format the NextJS i18n config uses, so it's perfectly fine to pass the same exact configuration. Just keep in mind that only `defaultLocale` and `locales` matter in this case because it's not like we can add subdomains to our apps.

The `locale` field is now populated in getProps() function calls with the current locale. It's also possible to switch locale either manually or via the new options object by passing it as a `locale` key.

Also, just for fun, I added support for Url-like formatting in the `fly()` function! uwu