feat(api): now possible to bring own logger
This both simplifies the package (11% of the size was pino) and allows for much greater customization. This can be super useful to external transports like Logtail!

Heck, it doesn't even have to be a Pino logger. Anything that matches the interface works fine~

I've also refactored pilot.ts log usage to use the `log` function for greater consistency and got rid of _getId() which felt dirty at the time. It had a nice effect of much cleaner code!

BREAKING CHANGE: `logLevel` config field was replaced by `logger`.