fix(pilot): better handling of errors caused by /500
It may be a bit awkward, but it's possible that the custom 500 error page itself could have errors. When that happens, we can't just recursively fall back to the same page forever...

SO! Let's just throw an actual error instead so the error can be more easily spotted and fixed!

As a side benefit, this also now guarantees that "load-complete" will always be called.