Modified create-next-app fork and templates
It now actually creates a PilotJS project!

There's still a lot of NextJS references, but that's okay. I'll be leaving a lot of these anyway, as Vercel definitely deserves credit for this. ~98% of this package is forked from them, after all.

There's also several more things that need to be done. For example, some dependencies don't get installed as devDependencies. The app.json remains unchanged, when it should at least change the name. Some dependencies here are also pointless for a basic template such as "next-build-id".