feat(pilot): support for build-time config
The CLI can now generate a default config based on a user-provided file that will be used in runtime.

This way, you can define certain things like host and i18n outside of runtime.

This now lives inside a _generated directory for consistency. _generated_pages.js will be deprecated and removed in v3.0.