feat(pilot): getProps functions now truly run on server-side
Calls to getServerSideProps and getStaticProps will now be handled in the server's Node runtime just like NextJS expects it to be.

This should now bring it to 100% parity in this regard and is the default as long as a "host" config is provided. The host is meant to be a URL where the web is hosted.

It will simply call the API handler provided (assuming it was applied correctly) to trigger those functions on the server side and delegate the result to the apps.

This means you can now use server-side secrets without exposing them to the app. The same goes for being able to use Node APIs such as "fs"!