fix(ci): fixed canary/next releases & separated api package
I forgot that the --tag also goes in the publish command and not just the version command. This should fix things so I don't accidentally set canary builds as "latest" again...

I've also separated the new /api package into it's own submodule just like Vercel does with next/link. This way we can avoid bundler issues with other modules, which is very important as we'll be using "fs" in /api which Metro is known to have issues with.