chore(pkg): reset private flag
Yarn gave warnings about nohoist not being respected as long as this flag was set to false, so I removed them. (true by default)

Turns out changesets is nice enough to respect these, so that wasn't quite the solution. Resetting these, yarn install now gives that warning away but it's not actually doing what it said it'd do, so that's a good sign...

It works again now at least!