feat: "dev" command will now auto detect package manager & allow custom commands
Yarn is no longer hardcoded, and it's now possible to specify your own commands in case you have a special use-case. In the case of WavePlay, there's already a need for this.

If your custom command doesn't specify a port, it will automatically append it to your `next dev` command.