feat: integrated cli into main package
I didn't think the CLI would become this important this quickly. It was initially intended to be pure mild convenience but now it's become an integral part of it.

It's still possible to use Pilot.js without the CLI, but it's no longer recommended now that web props are a thing.

I may switch over to PNPM soon to switch a lot of these dependencies over to devDependencies. Yarn seems to take issue with that due to hoisting.