feat: tunnel-less dev environment
Rather than rely on a tunnel for all connections, it's now optional and works the same way as Expo Go: using an external IPv4 address.

If a tunnel is detected, Pilot.js will prefer that, but it will use the local IP by default otherwise. As a bonus, start-up time is much faster now!

BREAKING: Tunnels are now opt-in rather than opt-out. This breaks assumptions for anyone that assumed otherwise.