docs: created mdx site (wip)
I decided to generate this using `create-pilot-app` because I was curious if it'd be possible to run this on native just for fun!

Turns out Metro has difficulties resolving these so... I'll have to link it to the mdx loader somehow or have to fork it and create a native version. Because, you know, why not?