fix: continue using next-transpile-modules due to issues with transpilePackages
I thought we were done using this dependency but guess I was wrong. :/

This has issues not transpiling .js files which is essential for the managed document. I'll look for a workaround later but so far transpiling this file ourselves isn't a solution due to it still importing `react-native` and the SWC plugin for RNW not going through dependencies.