refactor(cli): removed internal compiler as part of build process
This process was meant to be the more "reliable" method but turned out to be the least reliable because it tries to compile one file per page to get the context which simply doesn't work and ended up always using the fallback method. I mean, it did work, but only for extremely simple files...

At worst, it caused the build command to fail because of the compiler itself or the page code.

I have an idea to leverage Next.js itself in the future for this which will resolve a lot of issues. For now, we're only using the fallback method exclusively.